Update Live Site (#6)
* Changed Broncobots to Brookline Bots

* Created Meet the Team page

* Fix Site Breaking Changes

* Added Jekyll starter

* Mac only supper for .sh

* Move Start Jekyll scripts to folder

* Added confetti to rotary

added the option to click for confetti on the rotary sponsor square

* Added award details

included that we qualified for the premiere events

* Added Sponsor Us details - donations

updated the language to be accurate to our team and ways to donate

* created .yml file for members info

added 2 example people to the .yml file and updated some info on the branding page to change it to BBOTS info

* Changed color to 5777a8

* Added meet the team yml, html, md

* Meet the Team updated pages

* Captains

* adding sponsors

added Robert Rizika as a sponsor

* Added html and copied from sponsors

* Formatted SponsorUs w/ MarkDown

deleted bad stuff that was from bronco bots

* Changed Camp page Image

updated summer camp page

* Added Camp Image

I added the more cropped image

* Updated Camp Image

cropped the image to better fit the banner

* Darkened summer camp image

* Improved Summer Camp Image

edited it in light room

* Updated Summer camp Image

* No html in header

* Delete extra files

* Add team to default layout

Sorry Brielle :(

* No quoting names

* Changed format of team yml and html

* Fix mobile colors

* (WIP) Created mentor page

* Updated to say Brookline Bots on cover

* Update awards.md

* new startup

* Incorporated photos of mentors

* Upload Team pictures

* FOC Ignore

* Update start_jekyll.sh

* Update start_jekyll.sh

* (WIP) Put mentors in Meet The Team and added photos and roles

* Added home page image

* Added team image within description on home page

* Added awards photo

* Changed photos to more pro photos

* Changed Wiser's last name

* Added branding banner

* Changed headshot-less images

* Changed footers and team blurbs

* deleted most of the exess videos and fixed up the images on branding

* Create BrandingHeader.png

* Fixed Colors.png

* Update Branding.md

* Create AwardsHeader.jpg

* Changed based on comments

* Fixed spelling in Sponsor Us

* Deleted websiteTestImage

* Deleted websiteTestImage1

* Added illiustrator file and fixed brade color sizes; branding pretty much done

* Changed based on comments from PR 2.0

* Updated branding image

* Updated branding banner back to what it was

* Added Class of '(number)

* Changed based on comments (3)

* Matched photos

* Updated SponsorUs with accurate info

got rid of any info from matthew's high school or any online donation methods that we have no yet figured out

* Updated Images and Colors

I got rid of any of the broncobots stuff. I made a new image of all of the colors that we consistently are using.

* Added Sponsor Images and Organized Page

I added images for a bunch of the sponsors that did not already have images. I deleted sponsors that no longer work with us/ no longer exist as companies. I further organized the companies by activity with us/ most recent donations.

* Re-add BrandingColors

* Removed the alt & NoImage

i deleted the part of the code that has the alt= in response to a comment

* Removing online donation info

getting rid of the lines that talk about donating online and making brookline bots in quotes

* Readd HighRes.png and Logo.ai

* Added back HighRes & Illustrator versions

* Added back other logo types

* Fixed videos.html and removed 1987 video

* Added Adobe Icon & Deleted Bronco Bots videos

changed name of video in .yml, added adobe icon to branding, added adobe icon file.

* Got rid of broncobots video

* Fix png to svg

* Fixed Glassman and Warriors Images

replaced bad images with better ones in the sponsors.yml and sponsors images folder

* Removed BroncoBots images and replaced headers with BBOTs theme

replaced the headers of pages we have no quite yet finished with bbots themed header and got rid of some images that were exclusive to broncobots

* Updated Helpful Linnks with FTC & BBOTs rescources

Changed FRC specific resources to FTC things. Added images for Game Manual 0 as well as FTC scout with our teams page. Moved around the different links and resources by importance.

* Replaced Helpful Links Images

put correct images for each of the helpful links. Got rid of anything that we as a team do not use.

* Added Yael Photo

* remove Assets Page

removed the assets page because we did not have enough assets to actually fill it.

* Create Sophia.jpg

* Update Assets.yml

Hid the "past robots" tab/section

* Update sponsors.md

Changed the ugly header

* Deleted timeline and images

Deleted timeline and image from resources

* Fixed dates

made all of the dates the same format, resolved all of the comments on the PR (most not relevant)

* When I say don't use this image... I mean stop using this image.

* Fixed Header Image and Footer spacing

Also deleted unessesary files

* Changed 1 letter

Sponsor us => Sponsor Us

* Added and fixed wrong header

* How do we mess this up 4 times in a row

* deleted files not about our team

i got rid of all of the files that were about FRC stuff or bronco bots

* Fixed Image Headers

i updated image headers so that they would be our own photos

* Updated Sponsors Photo

i know this is not the best quality because i pulled it from the SignUp Genius

* Don't use low res images. the last one was fine but could be updated (but pick a good image. )

* Proper names are important

* Name files better

---------

Co-authored-by: BrooklineBots <148726436+BrooklineBots@users.noreply.github.com>
Co-authored-by: briellegithub <briellehwang@gmail.com>
Co-authored-by: Lila Cannon <26cannonl@brooklinek12.org>
Co-authored-by: Toliste <151093747+Toliste@users.noreply.github.com>